requirements.txt
streamlit==1.41.1
yfinance
newspaper3k
textblob
requests
openai
vaderSentiment
python-dotenv==0.21.1
#langchain.agents
langchain_core.tools 
langchain.llms
langchain_google_genai
pip>=24.3.1

